Pass picture data from camera activity to  image show activity for Room
Created a data class to store picture details like name, current date, uri and type of document selected in preview and made it parcelable to pass it from camera after capture to the activity that shows the image to prepare it for storing in Room. The option is selected using a spinner in preview of camera.